+++ io/include/io/reader.hpp
@@ -3,12 +3,12 @@
 #include <functional>
 #include <vecmem/memory/host_memory_resource.hpp>
 
-#include "csv/csv_io.hpp"
 #include "edm/cell.hpp"
 #include "edm/cluster.hpp"
 #include "edm/measurement.hpp"
 #include "edm/spacepoint.hpp"
-#include "tmp_edm.hpp"
+#include "io/csv.hpp"
+#include "io/demonstrator_edm.hpp"
 
 using namespace std::placeholders;
